requirements.txt
streamlit>=1.31.0
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
transformers>=4.39.0
torch>=2.1.0
pypdf>=4.0.0
numpy>=1.24.0
nltk>=3.9.1
pymupdf>=1.24.0
pandas>=2.1.0
